services/security-dashboard/requirements.txt
streamlit==1.50.0
requests
pandas
python-jose[cryptography]
typing-extensions>=4.14.0
altair==5.3.0